feat: 일정 삭제 시 비밀번호 검증 기능 추가 - 일정 삭제 전에 비밀번호 확인 절차 추가 - 올바른 비밀번호 입력 시에만 삭제 가능하도록 변경

+++ src/main/java/com/example/schedule/service/ScheduleServiceImpl.java
@@ -74,6 +74,13 @@ public class ScheduleServiceImpl implements ScheduleService{
 
     @Override
     public void deleteSchedule(Long id, String password) {
+
+        Schedule schedule = scheduleRepository.findScheduleByIDOrElseThrow(id);
+
+        if (!schedule.getPassword().equals(password)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "비밀번호가 틀렸습니다.");
+        }
+
         int deletedRow = scheduleRepository.deleteSchedule(id);
 
         if(deletedRow == 0) {
